product: move BUGZILLA_STATES to product module
The errata_tool_product module is the only module that needs this list,
so it makes sense to store this directly in the errata_tool_product
module instead of storing it in common_errata_tool.
